[Rename] auth 폴더 이동
